mbm: refactor - silo domain policy into silodomain
