+++ pkg/agent/qrm-plugins/mb/controller/policy/silodomain/mb_policy_constraint.go
@@ -14,14 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package policy
+package silodomain
 
 import (
 	"sync"
 
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/mbdomain"
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/plan"
-	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/qospolicy"
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/silodomain/qospolicy"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/strategy"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/monitor/stat"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/qosgroup"
@@ -57,7 +58,7 @@ func (c *constraintDomainMBPolicy) GetPlan(totalMB int, domain *mbdomain.MBDomai
 	return c.qosMBPolicy.GetPlan(totalMB, currQoSMB, qos, true)
 }
 
-func newConstraintDomainMBPolicy(qosMBPolicy qospolicy.QoSMBPolicy) DomainMBPolicy {
+func newConstraintDomainMBPolicy(qosMBPolicy qospolicy.QoSMBPolicy) policy.DomainMBPolicy {
 	return &constraintDomainMBPolicy{
 		qosMBPolicy: qosMBPolicy,
 	}
@@ -69,7 +70,7 @@ func newConstraintDomainMBPolicy(qosMBPolicy qospolicy.QoSMBPolicy) DomainMBPoli
 //	return NewConstraintDomainMBPolicy(ccdMBMin, strategy.ExtremeThrottle, strategy.HalfEase)
 //}
 
-func NewConstraintDomainMBPolicy(ccdMBMin int, throttleType, easeType strategy.LowPrioPlannerType) DomainMBPolicy {
+func NewConstraintDomainMBPolicy(ccdMBMin int, throttleType, easeType strategy.LowPrioPlannerType) policy.DomainMBPolicy {
 	qosMBPolicy := qospolicy.BuildHiPrioDetectedQoSMBPolicy(ccdMBMin, throttleType, easeType)
 	return newConstraintDomainMBPolicy(qosMBPolicy)
 }
