+++ pkg/agent/qrm-plugins/mb/controller/policy/silodomain/mb_policy_constraint_test.go
@@ -1,10 +1,10 @@
-package policy
+package silodomain
 
 import (
-	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/monitor/stat"
 	"reflect"
 	"testing"
 
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/monitor/stat"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/qosgroup"
 )
